[sensors] Several fixes in the Collision sensor
The collision sensor (as tested in Blender 2.74) was basically useless
since it would detect and return any collisions inside a 1x1x1m box
around the sensor. The unit-test was not checking this appropriately.

This commit scales down the collision bounding-box so that it matches
the mesh (a few cubic centimeters). It should fix #642.

While here:
- renamed the property that allows to select which objects are checked
  for collision from 'collision_property' to
  'only_objects_with_property'
- large documentation update, in particular through a useful and
  complete example.